# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MATCHPIT API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listMyBookings
      tags: [bookings]
      summary: Get my bookings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, past, cancelled]
      responses:
        "200":
          description: My bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a private turf booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Slot unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking detail
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}/cancel:
    post:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ─────────── HOSTED MATCHES ───────────
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Admin platform statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
  /admin/bookings:
    get:
      operationId: listAdminBookings
      tags: [admin]
      summary: List all bookings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /admin/hosted-matches:
    get:
      operationId: listAdminHostedMatches
      tags: [admin]
      summary: List all hosted matches
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Hosted matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostedMatch"
  /admin/payments:
    get:
      operationId: listAdminPayments
      tags: [admin]
      summary: Payment ledger
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
  /admin/venues:
    get:
      operationId: listAdminVenues
      tags: [admin]
      summary: List venues for approval
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VenueDetail"
  /admin/owner-leads:
    get:
      operationId: listAdminOwnerLeads
      tags: [admin]
      summary: List owner leads
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Owner leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OwnerLead"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    # ─── PROFILE ───
    Profile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        favoriteSports:
          type: array
          items:
            type: string
        avatarUrl:
          type: string
          nullable: true
        walletBalance:
          type: number
        badgeCount:
          type: integer
        trustScore:
          type: number
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, clerkId, fullName, email, walletBalance, badgeCount, trustScore, isAdmin, createdAt, favoriteSports]
...
    Venue:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string
        address:
          type: string
        sports:
          type: array
          items:
            type: string
        pricePerHour:
          type: number
        coverImage:
          type: string
          nullable: true
        rating:
          type: number
        totalReviews:
          type: integer
        isApproved:
          type: boolean
        amenities:
          type: array
          items:
            type: string
      required: [id, name, city, address, sports, pricePerHour, rating, totalReviews, isApproved, amenities]
    VenueDetail:
      allOf:
        - $ref: "#/components/schemas/Venue"
        - type: object
          properties:
            images:
              type: array
              items:
                type: string
            description:
              type: string
              nullable: true
            openTime:
              type: string
            closeTime:
              type: string
            contactPhone:
              type: string
              nullable: true
            ownerName:
              type: string
              nullable: true
            upcomingMatches:
              type: integer
          required: [images, openTime, closeTime, upcomingMatches]
...
    Booking:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        venueId:
          type: string
        slotId:
          type: string
        sport:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending_payment, confirmed, cancelled, completed]
        paymentId:
          type: string
          nullable: true
        venue:
          $ref: "#/components/schemas/Venue"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, venueId, slotId, sport, date, startTime, endTime, totalAmount, status, createdAt]
    CreateBookingBody:
      type: object
      properties:
        venueId:
          type: string
        slotId:
          type: string
        sport:
          type: string
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
        razorpaySignature:
          type: string
      required: [venueId, slotId, sport, razorpayOrderId, razorpayPaymentId, razorpaySignature]

    # ─── HOSTED MATCHES ───
    HostedMatch:
      type: object
      properties:
        id:
          type: string
        hostUserId:
          type: string
        venueId:
          type: string
        slotId:
          type: string
        sport:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        totalPlayers:
          type: integer
        minPlayers:
          type: integer
        currentPlayers:
          type: integer
        skillLevel:
          type: string
          enum: [beginner, intermediate, advanced, any]
        hostFee:
          type: number
        reserveFee:
          type: number
        finalFeePerPlayer:
          type: number
        totalVenueCost:
          type: number
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [open, confirmed, funded, cancelled, expired]
        financialStatus:
          type: string
          enum: [pending, partially_funded, funded]
        venue:
          $ref: "#/components/schemas/Venue"
          nullable: true
        host:
          $ref: "#/components/schemas/Profile"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, hostUserId, venueId, slotId, sport, date, startTime, endTime, totalPlayers, minPlayers, currentPlayers, skillLevel, hostFee, reserveFee, finalFeePerPlayer, totalVenueCost, status, financialStatus, createdAt]
...
    Payment:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [booking, host_commitment, match_reserve, match_final, refund, cashback]
        referenceId:
          type: string
          nullable: true
        razorpayOrderId:
          type: string
          nullable: true
        razorpayPaymentId:
          type: string
          nullable: true
        amount:
          type: number
        status:
          type: string
          enum: [pending, success, failed, refunded]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, amount, status, createdAt]

    # ─── WALLET ───
...
    OwnerLead:
      type: object
      properties:
        id:
          type: string
        venueName:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        city:
          type: string
        sports:
          type: array
          items:
            type: string
        message:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, contacted, onboarded, rejected]
        createdAt:
          type: string
          format: date-time
      required: [id, venueName, ownerName, phone, city, sports, status, createdAt]

    # ─── ADMIN ───
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalVenues:
          type: integer
        totalBookings:
          type: integer
        totalHostedMatches:
          type: integer
        totalRevenue:
          type: number
        activeMatches:
          type: integer
        pendingVenueApprovals:
          type: integer
        newOwnerLeads:
          type: integer
      required: [totalUsers, totalVenues, totalBookings, totalHostedMatches, totalRevenue, activeMatches, pendingVenueApprovals, newOwnerLeads]
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/Profile"
        total:
          type: integer
        page:
          type: integer
      required: [users, total, page]